pyproject.toml
[project]
name = "alfred-fpl"
version = "0.1.0"
description = "Alfred FPL — Fantasy Premier League BI agent domain"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    # Core engine (from PyPI) — import name is "alfred", not "alfredagain"
    "alfredagain>=2.0.0",

    # Database
    "supabase>=2.0.0",

    # Data / analytics (for Python executor in ANALYZE steps)
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pandera>=0.18.0",

    # HTTP (FPL API client)
    "requests>=2.31.0",
    "httpx>=0.27.0",

    # Environment
    "python-dotenv>=1.0.0",

    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/alfred_fpl"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
